# cmd/api/router.go
package main

import (
	"fmt"
	"net/http"
	"os"

	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	"github.com/gin-gonic/gin"
)

func (app *application) routes() http.Handler {
	g := gin.Default()

	v1 := g.Group("/api/v1")

	{
		v1.GET("/ping", ping)

		v1.POST("/auth/register", app.registerUser)
		v1.POST("/auth/login", app.login)

		v1.GET("/books/:id", app.getBook)
		v1.GET("/books", app.getPageOfBooks)
	}

	authGroup := v1.Group("/")
	authGroup.Use(app.AuthMiddleware())

	{
		authGroup.GET("/users/:id", app.getUser)
		authGroup.PUT("/users/:id", app.updateUser)
		authGroup.DELETE("/users/:id", app.deleteUser)

		authGroup.POST("/books", app.createBook)
		authGroup.PUT("/books/:id", app.updateBook)
		authGroup.DELETE("/books/:id", app.deleteBook)

		authGroup.GET("/orders", app.getPageOfOrders)
		authGroup.GET("/orders/:id", app.getOrder)
		authGroup.POST("/orders", app.createOrder)
		authGroup.PUT("/orders/:id", app.updateOrder)
		authGroup.DELETE("/orders/:id", app.deleteOrder)
	}

	v2 := g.Group("/api/v2")
	authGroup = v2.Group("/")
	authGroup.Use(app.AuthMiddleware())
	{
		authGroup.GET("/users/all", app.getAllUsers)
		authGroup.GET("/books/all", app.getAllBooks)
		authGroup.GET("/orders/all", app.getAllOrders)
	}

	g.GET("/swagger/*any", func(c *gin.Context) {
		if c.Request.RequestURI == "/swagger/" {
			c.Redirect(http.StatusFound, "/swagger/index.html")
		}
		url := fmt.Sprintf("http:%s:%s/swagger/doc.json", os.Getenv("DB_HOST"), os.Getenv("PORT"))
		ginSwagger.WrapHandler(swaggerFiles.Handler, ginSwagger.URL(url))(c)
	})

	return g
}

func ping(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"status": http.StatusOK, "version": "v1"})
}